requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
numpy
pydantic
jinja2
torch==2.2.0+cpu --extra-index-url https://download.pytorch.org/whl/cpu
